+++ pygeo/parameterization/DVGeoSketch.py
@@ -11,6 +11,8 @@ Enables the use of ESP (Engineering Sketch Pad) and OpenVSP (Open Vehicle Sketch
 # ======================================================================
 from abc import abstractmethod
 from collections import OrderedDict
+import numpy as np
+import copy
 from mpi4py import MPI
 from .BaseDVGeo import BaseDVGeometry
 from pyspline.utils import openTecplot, closeTecplot, writeTecplot1D
@@ -75,9 +77,6 @@ class DVGeoSketch(BaseDVGeometry):
         # Initial list of DVs
         self.DVs = OrderedDict()
 
-        self.useCompostiveDVs = False
-        self.compositeDVs = None
-
     def getValues(self):
         """
         Generic routine to return the current set of design
@@ -93,7 +92,7 @@ class DVGeoSketch(BaseDVGeometry):
         for dvName in self.DVs:
             dvDict[dvName] = self.DVs[dvName].value
 
-        if self.useCompostiveDVs:
+        if self.useComposite:
             dvDict = self.mapXDictToComp(dvDict)
 
         return dvDict
@@ -107,12 +106,178 @@ class DVGeoSketch(BaseDVGeometry):
         --------
         optProb.addCon(.....wrt=DVGeo.getVarNames())
         """
-        if  not self.useCompostiveDVs:
+        if not self.useComposite:
             names = list(self.DVs.keys())
         else:
-            names = [self.compositeDVs.name]
+            names = [self.DVComposite.name]
 
         return names
+
+    def mapXDictToDVGeo(self, inDict):
+        """
+        Map a dictionary of DVs to the 'DVGeo' design, while keeping non-DVGeo DVs in place
+        without modifying them
+        Parameters
+        ----------
+        inDict : dict
+            The dictionary of DVs to be mapped
+        Returns
+        -------
+        dict
+            The mapped DVs in the same dictionary format
+        """
+        # first make a copy so we don't modify in place
+
+        inDict = copy.deepcopy(inDict)
+        for key in inDict:
+            if key == self.DVComposite.name:
+                userVec = inDict[key]
+        # userVec = inDict[self.DVComposite.name]
+        outVec = self.mapVecToDVGeo(userVec)
+        outDict = self.convertSensitivityToDict(outVec.reshape(1, -1), out1D=True, useCompositeNames=False)
+        # now merge inDict and outDict
+        for key in inDict:
+            outDict[key] = inDict[key]
+        return outDict
+
+    def convertDictToSensitivity(self, dIdxDict):
+        """
+        This function performs the reverse operation of
+        convertSensitivityToDict(); it transforms the dictionary back
+        into an array. This function is important for the matrix-free
+        interface.
+        Parameters
+        ----------
+        dIdxDict : dictionary
+           Dictionary of information keyed by this object's
+           design variables
+        Returns
+        -------
+        dIdx : array
+           Flattened array of length getNDV().
+        """
+        DVCount = self.getNDV()
+
+        dIdx = np.zeros(DVCount, dtype="d")
+        i = 0
+        for key in self.DVs:
+            dv = self.DVs[key]
+            dIdx[i : i + dv.nVal] = dIdxDict[key]
+            i += dv.nVal
+
+        return dIdx
+
+    def mapXDictToComp(self, inDict):
+        """
+        The inverse of :func:`mapXDictToDVGeo`, where we map the DVs to the composite space
+        Parameters
+        ----------
+        inDict : dict
+            The DVs to be mapped
+        Returns
+        -------
+        dict
+            The mapped DVs
+        """
+        # first make a copy so we don't modify in place
+        inDict = copy.deepcopy(inDict)
+        userVec = self.convertDictToSensitivity(inDict)
+        outVec = self.mapVecToComp(userVec)
+        outDict = self.convertSensitivityToDict(outVec.reshape(1, -1), out1D=True, useCompositeNames=True)
+        return outDict
+
+    def mapVecToDVGeo(self, inVec):
+        """
+        This is the vector version of :func:`mapXDictToDVGeo`, where the actual mapping is done
+        Parameters
+        ----------
+        inVec : ndarray
+            The DVs in a single 1D array
+        Returns
+        -------
+        ndarray
+            The mapped DVs in a single 1D array
+        """
+        inVec = inVec.reshape(self.getNDV(), -1)
+        outVec = self.DVComposite.u @ inVec
+        return outVec.flatten()
+
+    def mapVecToComp(self, inVec):
+        """
+        This is the vector version of :func:`mapXDictToComp`, where the actual mapping is done
+        Parameters
+        ----------
+        inVec : ndarray
+            The DVs in a single 1D array
+        Returns
+        -------
+        ndarray
+            The mapped DVs in a single 1D array
+        """
+        inVec = inVec.reshape(self.getNDV(), -1)
+        outVec = self.DVComposite.u.transpose() @ inVec
+        return outVec.flatten()
+
+    def mapSensToComp(self, inVec):
+        """
+        Maps the sensitivity matrix to the composite design space
+        Parameters
+        ----------
+        inVec : ndarray
+            The sensitivities to be mapped
+        Returns
+        -------
+        ndarray
+            The mapped sensitivity matrix
+        """
+        outVec = inVec @ self.DVComposite.u  # this is the same as (self.DVComposite.u.T @ inVec.T).T
+        return outVec
+
+    def convertSensitivityToDict(self, dIdx, out1D=False, useCompositeNames=False):
+        """
+        This function takes the result of totalSensitivity and
+        converts it to a dict for use in pyOptSparse
+        Parameters
+        ----------
+        dIdx : array
+           Flattened array of length getNDV(). Generally it comes from
+           a call to totalSensitivity()
+        out1D : boolean
+            If true, creates a 1D array in the dictionary instead of 2D.
+            This function is used in the matrix-vector product calculation.
+        useCompositeNames : boolean
+            Whether the sensitivity dIdx is with respect to the composite DVs or the original DVGeo DVs.
+            If False, the returned dictionary will have keys corresponding to the original set of geometric DVs.
+            If True,  the returned dictionary will have replace those with a single key corresponding to the composite DV name.
+        Returns
+        -------
+        dIdxDict : dictionary
+           Dictionary of the same information keyed by this object's
+           design variables
+        """
+
+        # compute the various DV offsets
+        # DVCountGlobal= self._getDVOffsets()
+
+        i = 0
+        dIdxDict = {}
+        for key in self.DVs:
+            dv = self.DVs[key]
+            if out1D:
+                dIdxDict[key] = np.ravel(dIdx[:, i : i + dv.nVal])
+            else:
+                dIdxDict[key] = np.array(dIdx[:, i : i + dv.nVal])
+            i += dv.nVal
+
+        # replace other names with user
+        if useCompositeNames and self.useComposite:
+            array = []
+            for _key, val in dIdxDict.items():
+                array.append(val)
+            array = np.column_stack(array)
+            dIdxDict = {self.DVComposite.name: array}
+
+        return dIdxDict
 
     @abstractmethod
     def addVariable(self):
@@ -131,7 +296,7 @@ class DVGeoSketch(BaseDVGeometry):
             Optimization problem definition to which variables are added
         """
         # add the linear DV constraints that replace the existing bounds!
-        if self.useCompostiveDVs:
+        if self.useComposite:
             dv = self.DVComposite
             optProb.addVarGroup(dv.name, dv.nVal, "c", value=dv.value, lower=dv.lower, upper=dv.upper, scale=dv.scale)
             lb = {}
@@ -145,8 +310,8 @@ class DVGeoSketch(BaseDVGeometry):
             lb = self.convertDictToSensitivity(lb)
             ub = self.convertDictToSensitivity(ub)
 
-            # self.compositeDVs.lower=lb
-            # self.compositeDVs.upper=ub
+            # self.DVComposite.lower=lb
+            # self.DVComposite.upper=ub
 
             optProb.addConGroup(
                 f"{self.DVComposite.name}_con",
@@ -161,7 +326,9 @@ class DVGeoSketch(BaseDVGeometry):
         else:
             for dvName in self.DVs:
                 dv = self.DVs[dvName]
-                optProb.addVarGroup(dvName, dv.nVal, "c", value=dv.value, lower=dv.lower, upper=dv.upper, scale=dv.scale)
+                optProb.addVarGroup(
+                    dvName, dv.nVal, "c", value=dv.value, lower=dv.lower, upper=dv.upper, scale=dv.scale
+                )
 
     def writePointSet(self, name, fileName):
         """
